Performance: use FNV hashmap for the dictionary words
The default HashMap is built to withstand DoS attacks. We don't have
to worry about when loading a dictionary so we can use a faster hash.
Adding user-supplied words might be an attack vector, but the user
normally has control over which words are added. (And if not, that's
already a problem.)

This change gave a 30% performance improvement for the en_US benchmark.